requirements.txt
streamlit
google-generativeai
python-dotenv
lanchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community
